requirements.txt
streamlit
pypdf
sentence-transformers
chromadb
numpy
torch
python-dotenv
groq
python-dotenv